fix(test): address review findings — deduplicate tests, fix naming, extract fixtures (NIU-455)
